fix s3_stubber issue
