+++ tests/test_echo10_construction.py
@@ -1,7 +1,17 @@
 import json
 import unittest.mock
 
+import pytest
+from botocore.stub import Stubber
+
 import main
+
+
+@pytest.fixture
+def s3_stubber():
+    with Stubber(main.s3.meta.client) as stubber:
+        yield stubber
+        stubber.assert_no_pending_responses()
 
 
 def test_get_file_content_from_s3(inputs, test_data_dir, s3_stubber):
